Add files via upload
Searched for a database that contained all the street names in Chicago, IL and found a csv file.  
Ran a little project in a Jupiter notebook to answer the question: "How many streets in Chicago end with the suffix "dale"?

Note: I used to live on "Lawndale" in Evanston, IL.  Recently, while driving with my 7 year-old, he noticed a street named "Avondale" and then another named "Kingsdale".  He asked me if I knew how many streets in Chicago had the "dale" ending in their name, which inspired this little project!